extract page-data helpers used in page engine to separate module, to not bundle unneeded pieces
